Fix bug

+++ src/main/java/science/icebreaker/account/AccountProfileRepository.java
@@ -3,6 +3,6 @@ package science.icebreaker.account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-interface AccountProfileRepository extends JpaRepository<AccountProfile, Long> {
+interface AccountProfileRepository extends JpaRepository<AccountProfile, Integer> {
 
 }
