+++ src/main/java/science/icebreaker/account/AccountRoleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 
-public interface AccountRoleRepository extends JpaRepository<AccountRole, Long> {
+public interface AccountRoleRepository extends JpaRepository<AccountRole, Integer> {
 
     List<AccountRole> findAllByEmail(String email);
